Size of

+++ 0x00-hello_world/6-size.c
@@ -2,6 +2,8 @@
 
 /**
  *  main - Entry point
+ *
+ *  Return: Always 0(Success) 
  */
 
 int main(void)
@@ -13,10 +15,10 @@ int main(void)
 	long long int e;
 
 
-	printf("The size of an integer: %i byte(s)\n", (int)sizeof(a));
-	printf("The size of a float: %i byte(s)\n", (int)sizeof(b));
-	printf("The size of characters: %i byte(s)\n", (int)sizeof(c));
-	printf("The size of long integer: %i byte(s)\n", (int)sizeof(d));
-	printf("The size of a long long integer: %i byte(s)", (int)sizeof(e));
+	printf("The size of an integer: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("The size of a float: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("The size of characters: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("The size of long integer: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("The size of a long long integer: %lu byte(s)", (unsigned long)sizeof(e));
 	return (0);
 }
